# Conflicts:
#	backend/src/main/java/com/nocountry/backend/config/DefaultAdminRunner.java
#	backend/src/main/java/com/nocountry/backend/controller/AuthController.java
#	backend/src/main/java/com/nocountry/backend/controller/BookingController.java
#	backend/src/main/java/com/nocountry/backend/controller/CategoryController.java
#	backend/src/main/java/com/nocountry/backend/controller/LocationController.java
#	backend/src/main/java/com/nocountry/backend/dto/CarDto.java
#	backend/src/main/java/com/nocountry/backend/dto/CategoryDto.java
#	backend/src/main/java/com/nocountry/backend/dto/InformationLocationDto.java
#	backend/src/main/java/com/nocountry/backend/model/Location.java
#	backend/src/main/java/com/nocountry/backend/service/ILocationsService.java
#	backend/src/main/java/com/nocountry/backend/service/impl/BookingServiceImpl.java
#	backend/src/main/java/com/nocountry/backend/service/impl/BranchServiceImplI.java
#	backend/src/main/java/com/nocountry/backend/service/impl/CarServiceImpl.java
#	backend/src/main/java/com/nocountry/backend/service/impl/CustomerServiceImpl.java
#	backend/src/main/java/com/nocountry/backend/service/impl/LocationServiceImpl.java
#	backend/src/main/java/com/nocountry/backend/util/georefapi/impl/ExecuteImpl.java
#	backend/src/main/resources/data.sql

+++ backend/src/main/java/com/nocountry/backend/controller/CategoryController.java
@@ -3,7 +3,6 @@ package com.nocountry.backend.controller;
 import java.util.List;
 import java.util.Optional;
 
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -18,48 +17,59 @@ import org.springframework.web.bind.annotation.RestController;
 import com.nocountry.backend.dto.CategoryDto;
 import com.nocountry.backend.service.ICategoryService;
 
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 
 @RestController
 @RequestMapping("/api/v1/categories")
 @RequiredArgsConstructor
+@Tag(name = "Categories", description = "Management of vehicle categories in MoveAr. It allows creating, modifying, and deleting categories, as well as obtaining detailed information about them.")
 @SecurityRequirement(name = "bearerAuth")
 public class CategoryController {
 
     private final ICategoryService categoryService;
 
     @GetMapping("/all")
-    public ResponseEntity<List<CategoryDto>> getAllCategories() {
-        List<CategoryDto> categories = categoryService.findAllCategories();
+    @Operation(summary = "Get all Categories.")
+    public ResponseEntity<List<CategoryDto>> getAllCategorys() {
+        var categories = categoryService.findAllCategories();
         if (categories.isEmpty()) {
-            return ResponseEntity.noContent().build();
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } else {
-            return ResponseEntity.ok(categories);
+            return new ResponseEntity<>(categories, HttpStatus.OK);
         }
     }
 
     @GetMapping("/{categoryId}")
-    public ResponseEntity<CategoryDto> getCategory(@PathVariable(value = "categoryId") Long categoryId) {
-        CategoryDto categoryDto = categoryService.findCategoryById(categoryId);
-        return ResponseEntity.ok(categoryDto);
+    @Operation(summary = "Get a category by Id.")
+    public ResponseEntity<CategoryDto> getCategoryById(@PathVariable Long categoryId) {
+        return new ResponseEntity<>(categoryService.findCategoryById(categoryId), HttpStatus.OK);
     }
 
     @PostMapping("/create")
-    public ResponseEntity<CategoryDto> createCategory(@RequestBody CategoryDto categoryDto) {
-        CategoryDto categorySaved = categoryService.saveCategory(categoryDto);
-        return ResponseEntity.ok(categorySaved);
+    @Operation(summary = "Create a new category.")
+    public ResponseEntity<CategoryDto> createCategory(@RequestBody CategoryDto category) {
+        return new ResponseEntity<>(categoryService.saveCategory(category), HttpStatus.OK);
     }
 
     @PutMapping("/{categoryId}/update")
-    public ResponseEntity<CategoryDto> updateCategory(@PathVariable(value = "categoryId") Long categoryId,
-            @RequestBody CategoryDto categoryDto) {
-        CategoryDto saved = categoryService.updateCategory(categoryId, categoryDto);
-        return ResponseEntity.ok(saved);
+    @Operation(summary = "Update an existing category by Id.")
+    public ResponseEntity<CategoryDto> updateCategory(@PathVariable Long categoryId,
+            @RequestBody CategoryDto category) {
+        return new ResponseEntity<>(categoryService.updateCategory(categoryId, category), HttpStatus.OK);
     }
 
     @DeleteMapping("/{categoryId}/delete")
-    public ResponseEntity<String> deleteCategory(@PathVariable(value = "categoryId") Long categoryId) {
-        categoryService.deleteCategory(categoryId);
-        return new ResponseEntity<>("Category successfully deleted", HttpStatus.ACCEPTED);
+    @Operation(summary = "Delete an existing category by Id.")
+    public ResponseEntity<String> deleteCategory(@PathVariable Long categoryId) {
+        Optional<CategoryDto> category = Optional.ofNullable(categoryService.findCategoryById(categoryId));
+        if (category.isPresent()) {
+            categoryService.deleteCategory(categoryId);
+            return new ResponseEntity<>("Category successfully deleted", HttpStatus.ACCEPTED);
+        } else {
+            return new ResponseEntity<>("Category not found", HttpStatus.NOT_FOUND);
+        }
     }
 }
